+++ Sources/VComponents/Components/Modals/Modal/VModalUIModel.swift
@@ -42,14 +42,11 @@ public struct VModalUIModel {
     )
 
     // MARK: Properties - Corners
-    /// Rounded corners. Set to to `allCorners`.
-    public var roundedCorners: RectCorner = .allCorners
+    /// Corner radii . Set to to `15`s.
+    public var cornerRadii: RectangleCornerRadii = .init(15)
 
     /// Indicates if left and right corners should switch to support RTL languages. Set to `true`.
     public var reversesLeftAndRightCornersForRTLLanguages: Bool = true
-
-    /// Corner radius. Set to `15`.
-    public var cornerRadius: CGFloat = 15
 
     // MARK: Properties - Background
     /// Background color.
@@ -64,9 +61,8 @@ public struct VModalUIModel {
     var groupBoxSubUIModel: VGroupBoxUIModel {
         var uiModel: VGroupBoxUIModel = .init()
 
-        uiModel.roundedCorners = roundedCorners
+        uiModel.cornerRadii = cornerRadii
         uiModel.reversesLeftAndRightCornersForRTLLanguages = reversesLeftAndRightCornersForRTLLanguages
-        uiModel.cornerRadius = cornerRadius
 
         uiModel.backgroundColor = backgroundColor
 
@@ -172,8 +168,8 @@ extension VModalUIModel {
     public static var insettedContent: Self {
         var uiModel: Self = .init()
         
-        uiModel.contentMargins = Margins(uiModel.cornerRadius)
-        
+        uiModel.contentMargins = Margins(15)
+
         return uiModel
     }
 }