+++ Sources/VComponents/Components/Modals/SideBar/VSideBar.swift
@@ -114,7 +114,15 @@ struct VSideBar<Content>: View where Content: View {
             maxWidth: currentWidth,
             maxHeight: currentHeight
         )
-        .cornerRadius(uiModel.cornerRadius, corners: uiModel.roundedCorners) // Fixes issue of content-clipping, as it's not in `VGroupBox`
+        .clipShape( // Fixes issue of content-clipping, as it's not in `VGroupBox`
+            .rect(
+                cornerRadii: uiModel.cornerRadii
+                    .withReversedLeftAndRightCorners(
+                        uiModel.reversesLeftAndRightCornersForRTLLanguages &&
+                        layoutDirection.isRightToLeft
+                    )
+            )
+        )
         .offset(isPresentedInternally ? .zero : initialOffset)
         .gesture(
             DragGesture(minimumDistance: 20) // Non-zero value prevents collision with scrolling